Temporary switch to Float datatype for price && implemented error handling

+++ models/test.py
@@ -1,5 +1,5 @@
 import uuid
-from sqlalchemy import Column, String, DECIMAL
+from sqlalchemy import Column, String, Float, DECIMAL
 from sqlalchemy.ext.declarative import declarative_base
 
 Base = declarative_base()
@@ -10,6 +10,6 @@ class Test(Base):
     id = Column(String(36), primary_key=True, default=lambda: str(uuid.uuid4()))
     name = Column(String(50), index=True, unique=True)
     description = Column(String(255))
-    price  = Column(DECIMAL)
+    price  = Column(Float(4))
